refactor(core/server): extract request context into ./server/context.ts
First of several extractions out of the 1912-line `server.ts`. Moves
the request context module-scoped globals (`currentContext`,
`currentFilePath`, layout WeakMap) and every function that reads or
writes them into a dedicated module:

- `setRequestContext`, `getRequestContext`, `clearRequestContext`
- `setCurrentFilePath`, `getCurrentFilePath`
- `setActiveLayouts`, `getActiveLayouts`
- `getRequestEvent`
- `LayoutComponent` type

`server.ts` re-exports every symbol from `./server/context.ts` so
external imports against `@ts-api-kit/core` / `@ts-api-kit/core/server`
keep working unchanged. Internal consumers in `server.ts` (the
`jsxStream` helper and the handler's `finally` cleanup) were updated
to call through `getRequestContext()` / `clearRequestContext()` — they
no longer reach into module-scoped globals.

`wc -l`: server.ts 1912 → 1833; new context.ts is 118 lines. All
tests and type-check remain green.
